+++ data/src/main/java/com/example/data/di/NetworkModule.kt
@@ -2,9 +2,7 @@ package com.example.data.di
 
 import com.example.data.common.BASE_URL
 import com.example.data.common.TIME_OUT_POLICY
-import com.example.data.remote.FollowerService
-import com.example.data.remote.RepoService
-import com.example.data.remote.UserService
+import com.example.data.remote.service.GitHubApiService
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
@@ -20,7 +18,7 @@ import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
-object NetworkModule {
+class NetworkModule {
 
     @Singleton
     @Provides
@@ -71,19 +69,6 @@ object NetworkModule {
 
     @Singleton
     @Provides
-    fun provideGitHubRepoService(retrofit: Retrofit): RepoService {
-        return retrofit.create(RepoService::class.java)
-    }
-
-    @Singleton
-    @Provides
-    fun provideUserService(retrofit: Retrofit): FollowerService {
-        return retrofit.create(FollowerService::class.java)
-    }
-
-    @Singleton
-    @Provides
-    fun provideFollowerService(retrofit: Retrofit): UserService {
-        return retrofit.create(UserService::class.java)
-    }
+    fun provideGitHubService(retrofit: Retrofit): GitHubApiService =
+        retrofit.create(GitHubApiService::class.java)
 }
