--- data/src/main/java/com/example/data/remote/FollowerService.kt
@@ -1,12 +1,0 @@
-package com.example.data.remote
-
-import com.example.data.remote.response.follower.FollowerResponse
-import retrofit2.http.GET
-import retrofit2.http.Path
-
-interface FollowerService {
-    @GET("/users/{username}/followers")
-    suspend fun getFollowers(
-        @Path("username") userName: String,
-    ): List<FollowerResponse>
-}
