+++ data/src/main/java/com/example/data/remote/service/GitHubApiService.kt
@@ -1,0 +1,25 @@
+package com.example.data.remote.service
+
+import com.example.data.remote.response.follower.FollowerResponse
+import com.example.data.remote.response.repo.GitHubRepoResponse
+import com.example.data.remote.response.user.UserResponse
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface GitHubApiService {
+
+    @GET("/users/{username}/followers")
+    suspend fun getFollowers(
+        @Path("username") userName: String,
+    ): List<FollowerResponse>
+
+    @GET("/users/{username}/repos")
+    suspend fun getGitHubRepo(
+        @Path("username") userName: String,
+    ): List<GitHubRepoResponse>
+
+    @GET("/users/{username}")
+    suspend fun getUser(
+        @Path("username") userName: String,
+    ): UserResponse
+}
